Commit changed tests in a distinct commit when passing/passing builds (#584)
This PR intends to fix the issue #490.
A branch generated with the changes contained in this PR can be seen here: https://github.com/fermadeiral/bears-collection-pp/tree/INRIA-spoon-191511078-20170113-031556
The process is right (of committing changed tests in a distinct commit when passing/passing builds), but this branch, for instance, is not completely correct because of the issue #581. However, the two issues are different, thus I propose this PR for adding the feature for issue #490, and in another moment (PR) I will fix the bug reported in issue #581.

PS: One of the last commits is a bug fix (I introduced such bug when I extracted the method `copyDirectory`). Sorry for mixing it in this PR, but I needed to fix that to see if the things were working for the feature added in this PR.