Improve mail notification for end of process (#678)
fix #546 
Attempt at resolving issue #546. 

If I have understood the end of process notification correctly, then the start and end of the repair attempt was already mentioned in the mail. The only thing lacking was the duration, or if duration is not the same as uptime I have made a mistake. Should I perhaps `config.getInstance().getDuration().toString()` instead?