Changed the folder name associated with a build created by Astor in its workspace
Repairnator now gives the build id and the selected repair mode to Astor, so that Astor can create a unique folder for every build in the workspace. In this way, there are no conflicts in the names of the folders and it is possible to know the repair mode used by Astor to create the Program Variants contained in a specific folder.

The problem didn't occur with all the builds, but only with some of them (e.g., this one: 374796850).

Example of the old name: AstorMain-ExperimentDesigner
Example of the new name: AstorMain-AstorJKali-374796850

Signed-off-by: ginellidavide@gmail.com <ginellidavide@gmail.com>