Changed the root folder associated with a build in the workspace
For every build, Repairnator now creates a folder whose name is the build id, and clones in that folder the repository associated with that particular build id. In this way, every build has its proper folder in the workspace, without any conflict in the names of the folders.

Example of the old structure: workspace/surli/failingProject
Example of the new structure: workspace/564711868/surli/failingProject

Signed-off-by: ginellidavide@gmail.com <ginellidavide@gmail.com>